mysql_zero_downtime/scripts/requirements.txt
mysql-connector-python==8.2.0
Faker==19.12.0
